+++ test/test_malloc.cpp
@@ -20,7 +20,7 @@ namespace bc = boost::compute;
 
 BOOST_AUTO_TEST_CASE(malloc_int)
 {
-    bc::device_ptr<int> ptr = bc::experimental::malloc<int>(5);
+    bc::experimental::device_ptr<int> ptr = bc::experimental::malloc<int>(5);
 
     int input_data[] = { 2, 5, 8, 3, 6 };
     bc::copy(input_data, input_data + 5, ptr);
